Separated methods into GET and POST tied together by a parent class
