+++ Binance/info.txt
@@ -1,1 +1,1 @@
-{"symbol": "SHIBUSDT", "status": "TRADING", "baseAsset": "SHIB", "baseAssetPrecision": 2, "quoteAsset": "USDT", "quotePrecision": 8, "quoteAssetPrecision": 8, "baseCommissionPrecision": 2, "quoteCommissionPrecision": 8, "orderTypes": ["LIMIT", "LIMIT_MAKER", "MARKET", "STOP_LOSS_LIMIT", "TAKE_PROFIT_LIMIT"], "icebergAllowed": true, "ocoAllowed": true, "quoteOrderQtyMarketAllowed": true, "allowTrailingStop": true, "cancelReplaceAllowed": true, "isSpotTradingAllowed": true, "isMarginTradingAllowed": false, "filters": [{"filterType": "PRICE_FILTER", "minPrice": "0.00000001", "maxPrice": "1.00000000", "tickSize": "0.00000001"}, {"filterType": "PERCENT_PRICE", "multiplierUp": "5", "multiplierDown": "0.2", "avgPriceMins": 5}, {"filterType": "LOT_SIZE", "minQty": "1.00", "maxQty": "46116860414.00", "stepSize": "1.00"}, {"filterType": "MIN_NOTIONAL", "minNotional": "1.00000000", "applyToMarket": true, "avgPriceMins": 5}, {"filterType": "ICEBERG_PARTS", "limit": 10}, {"filterType": "MARKET_LOT_SIZE", "minQty": "0.00", "maxQty": "9445281885.47", "stepSize": "0.00"}, {"filterType": "TRAILING_DELTA", "minTrailingAboveDelta": 10, "maxTrailingAboveDelta": 2000, "minTrailingBelowDelta": 10, "maxTrailingBelowDelta": 2000}, {"filterType": "PERCENT_PRICE_BY_SIDE", "bidMultiplierUp": "1.25", "bidMultiplierDown": "0.2", "askMultiplierUp": "5", "askMultiplierDown": "0.75", "avgPriceMins": 5}, {"filterType": "MAX_NUM_ORDERS", "maxNumOrders": 200}, {"filterType": "MAX_NUM_ALGO_ORDERS", "maxNumAlgoOrders": 5}], "permissions": ["SPOT"], "defaultSelfTradePreventionMode": "EXPIRE_MAKER", "allowedSelfTradePreventionModes": ["EXPIRE_TAKER", "EXPIRE_MAKER", "EXPIRE_BOTH"]}
+{"symbol": "SHIBUSDT", "status": "TRADING", "baseAsset": "SHIB", "baseAssetPrecision": 2, "quoteAsset": "USDT", "quotePrecision": 8, "quoteAssetPrecision": 8, "baseCommissionPrecision": 2, "quoteCommissionPrecision": 8, "orderTypes": ["LIMIT", "LIMIT_MAKER", "MARKET", "STOP_LOSS_LIMIT", "TAKE_PROFIT_LIMIT"], "icebergAllowed": true, "ocoAllowed": true, "quoteOrderQtyMarketAllowed": true, "allowTrailingStop": true, "cancelReplaceAllowed": true, "isSpotTradingAllowed": true, "isMarginTradingAllowed": false, "filters": [{"filterType": "PRICE_FILTER", "minPrice": "0.00000001", "maxPrice": "1.00000000", "tickSize": "0.00000001"}, {"filterType": "PERCENT_PRICE", "multiplierUp": "5", "multiplierDown": "0.2", "avgPriceMins": 5}, {"filterType": "LOT_SIZE", "minQty": "1.00", "maxQty": "46116860414.00", "stepSize": "1.00"}, {"filterType": "MIN_NOTIONAL", "minNotional": "1.00000000", "applyToMarket": true, "avgPriceMins": 5}, {"filterType": "ICEBERG_PARTS", "limit": 10}, {"filterType": "MARKET_LOT_SIZE", "minQty": "0.00", "maxQty": "5178615062.90", "stepSize": "0.00"}, {"filterType": "TRAILING_DELTA", "minTrailingAboveDelta": 10, "maxTrailingAboveDelta": 2000, "minTrailingBelowDelta": 10, "maxTrailingBelowDelta": 2000}, {"filterType": "PERCENT_PRICE_BY_SIDE", "bidMultiplierUp": "1.25", "bidMultiplierDown": "0.2", "askMultiplierUp": "5", "askMultiplierDown": "0.75", "avgPriceMins": 5}, {"filterType": "MAX_NUM_ORDERS", "maxNumOrders": 200}, {"filterType": "MAX_NUM_ALGO_ORDERS", "maxNumAlgoOrders": 5}], "permissions": ["SPOT"], "defaultSelfTradePreventionMode": "EXPIRE_MAKER", "allowedSelfTradePreventionModes": ["EXPIRE_TAKER", "EXPIRE_MAKER", "EXPIRE_BOTH"]}
